--- lib/core/utils/constants/api_constants.dart
@@ -1,17 +1,0 @@
-abstract class ApiConstants {
-  static const String baseUrl = 'https://flower.elevateegy.com/api/v1/';
-  static const String signIn = 'auth/signin';
-  static const String token = 'token';
-  static const String contentType = 'Content-Type';
-  static const String applicationJson = 'application/json';
-  static const String bearer = 'Bearer';
-  static const String authorization = 'Authorization';
-  static const String occasions = 'occasions';
-  static const String products = 'products';
-  static const String occasionQuery = 'occasion';
-  static const String allCategories = 'categories';
-  static const String getProducts = 'products';
-  static const String categoryQuery = 'category';
-  static const String mostSelling = 'best-seller';
-  static const String orders = 'orders';
-}
